# Back-End/Api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phone Tracker API
  description: |
    Phone Tracker API for academic misconduct monitoring and student management.
    
    This API provides comprehensive endpoints for:
    - User authentication and session management
    - Student and system user management
    - Occurrence tracking and feedback systems
    - Analytics and reporting features
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Permissions
    The API uses role-based access control with the following roles:
    - **ADMIN**: Full system access
    - **STAFF**: Management operations
    - **STUDENT**: Limited access to own data
    - **TEACHER**: Educational tools access
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
    All error responses follow a consistent structure with timestamp, status, error type, message, and path.

  version: 1.0.0
  contact:
    name: Luiz Gustavo
    email: luiz.gustavo@aluno.ifsp.edu.br
    url: https://github.com/luizgustavosgobi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    OccurrenceDto:
      type: object
      required:
        - date
        - type
        - cameraId
      properties:
        date:
          type: string
          format: date-time
          description: When the occurrence happened (ISO 8601 format)
          example: "2024-01-15T14:30:00.000Z"
        type:
          type: string
          enum: [PHONE_USAGE, SLEEPING, TALKING, EATING, OTHER]
          description: |
            Type of misconduct detected:
            - PHONE_USAGE: Student using mobile device inappropriately
            - SLEEPING: Student sleeping during class
            - TALKING: Disruptive talking during class
            - EATING: Eating during prohibited times
            - OTHER: Other policy violations
          example: "PHONE_USAGE"
        cameraId:
          type: integer
          description: Numeric identifier of the camera that detected the occurrence
          minimum: 1
          maximum: 9999
          example: 1
        local:
          type: string
          description: Location where the occurrence happened (optional)
          maxLength: 100
          example: "Classroom A"
        proof:
          type: string
          format: binary
          description: Proof file (image/video evidence, max 10MB)
          maxLength: 10485760
      description: Occurrence data submitted by external detection systems
    OccurrenceResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique occurrence identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        local:
          type: string
          description: Location where the occurrence happened
          example: "Classroom A"
        date:
          type: string
          format: date-time
          description: When the occurrence happened (UTC)
          example: "2024-01-15T14:30:00.000Z"
        type:
          type: string
          enum: [PHONE_USAGE, SLEEPING, TALKING, EATING, OTHER]
          description: Type of misconduct detected
          example: "PHONE_USAGE"
        cameraId:
          type: string
          description: Camera identifier that detected the occurrence
          example: "CAM_001"
        proofFile:
          type: string
          description: Filename of the proof file (null if no proof available)
          nullable: true
          example: "proof_20240115_143000.jpg"
        confidence:
          type: number
          format: float
          description: AI confidence level (0.0 to 1.0, null if not AI-detected)
          minimum: 0.0
          maximum: 1.0
          nullable: true
          example: 0.95
        feedback:
          $ref: '#/components/schemas/FeedbackResponseDto'
      description: Complete occurrence information with optional feedback
    FeedbackDto:
      type: object
      required:
        - status
      properties:
        studentId:
          type: string
          description: Student ID to associate with approved occurrence (required for APPROVED status)
          minLength: 3
          maxLength: 20
          example: "20240001"
        comments:
          type: string
          description: Staff comments about the occurrence review (optional)
          maxLength: 1000
          example: "Valid occurrence confirmed after manual review"
        status:
          type: string
          enum: [APPROVED, REJECTED]
          description: |
            Review decision:
            - APPROVED: Occurrence is valid and will be associated with the student
            - REJECTED: Occurrence is invalid (false positive, etc.)
          example: "APPROVED"
      description: Staff feedback and review decision for an occurrence
    FeedbackResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique feedback identifier
          example: "456e7890-f12a-34b5-c678-901234567890"
        comments:
          type: string
          description: Staff comments about the occurrence
          nullable: true
          example: "Valid occurrence confirmed"
        status:
          type: string
          enum: [APPROVED, REJECTED, PENDING]
          description: Review status
          example: "APPROVED"
        reviewedBy:
          type: string
          description: ID of staff member who provided the feedback
          nullable: true
          example: "staff001"
        reviewDate:
          type: string
          format: date-time
          description: When the feedback was submitted (UTC)
          nullable: true
          example: "2024-01-15T15:00:00.000Z"
        student:
          $ref: '#/components/schemas/StudentSummaryDto'
      description: Complete feedback information with reviewer details
    StudentSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: Student unique identifier
          example: "20240001"
        name:
          type: string
          description: Student full name
          example: "Alice Johnson"
      description: Basic student information for feedback associations
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (UTC)
          example: "2024-01-01T12:00:00.000Z"
        status:
          type: integer
          description: HTTP status code
          minimum: 400
          maximum: 599
          example: 400
        error:
          type: string
          description: HTTP status reason phrase
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message explaining what went wrong
          example: "Validation failed for field 'name': must not be blank"
        path:
          type: string
          description: API endpoint that generated the error
          example: "/users/create"
        details:
          type: object
          description: Additional error details (validation errors, etc.)
          nullable: true
          additionalProperties: true
          example:
            field: "name"
            rejectedValue: ""
            code: "NotBlank"
      description: Standard error response format used across all endpoints
...
paths:
...
  /occurrences:
    post:
      tags:
        - Occurrences
      summary: Create a new occurrence (External API)
      description: |
        Creates a new occurrence in the system. This endpoint is designed for external systems 
        such as computer vision services and does not require user authentication.
        
        **Required Permissions:** None - Public endpoint for external systems
        
        **Use Cases:**
        - Automated detection systems reporting incidents
        - Camera systems with AI analysis
        - IoT devices detecting policy violations
        
        **Features:**
        - Automatic WebSocket notification to connected clients
        - Optional proof file upload with secure storage
        - Immediate availability for staff review
      operationId: createOccurrence
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OccurrenceDto'
      responses:
        '200':
          description: Occurrence created successfully and notifications sent
        '400':
          description: Invalid input data or proof file issues
        '500':
          description: Server error - Failed to save proof file or process occurrence

    get:
      tags:
        - Occurrences
      summary: Retrieve all occurrences with filtering
      description: |
        Retrieves a paginated list of all occurrences in the system with comprehensive filtering options.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Features:**
        - Advanced filtering by date range, student, type, location, camera
        - Paginated results with configurable sorting
        - Performance optimized for large datasets
        - Real-time data with latest occurrences
        
        **Default Sorting:** By date (newest first)
      operationId: getAllOccurrences
      parameters:
        - name: studentId
          in: query
          description: Filter by specific student ID
          schema:
            type: string
          example: "20240001"
        - name: type
          in: query
          description: Filter by occurrence type
          schema:
            type: string
          example: "PHONE_USAGE"
        - name: startDate
          in: query
          description: Filter occurrences from this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: endDate
          in: query
          description: Filter occurrences until this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
        - name: location
          in: query
          description: Filter by occurrence location
          schema:
            type: string
          example: "Classroom A"
        - name: camera
          in: query
          description: Filter by camera identifier
          schema:
            type: string
          example: "CAM_001"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            default: "date,desc"
      responses:
        '200':
          description: Occurrences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/OccurrenceResponseDto'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '400':
          description: Invalid filter parameters
        '401':
          description: Authentication required
        '403':
          description: Access denied
  /occurrences/{id}:
    get:
      tags:
        - Occurrences
      summary: Retrieve occurrence details by ID
      description: |
        Retrieves comprehensive information about a specific occurrence including proof files and feedback.
        
        **Required Permissions:** 
        - `ROLE_ADMIN` - Can view any occurrence
        - `ROLE_STAFF` - Can view any occurrence
        - Students can only view their own occurrences
        
        **Data Returned:** Complete occurrence details including proof files, feedback, and metadata.
      operationId: getOccurrence
      parameters:
        - name: id
          in: path
          required: true
          description: Unique UUID identifier of the occurrence
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Occurrence details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccurrenceResponseDto'
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Occurrence not found
  /occurrences/waiting-feedback:
    get:
      tags:
        - Occurrences
      summary: Retrieve occurrences awaiting feedback
      description: |
        Retrieves all occurrences that have not yet received feedback/review from staff.
        This endpoint is optimized for staff workflows and review queues.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Use Cases:**
        - Staff review dashboard
        - Automated workflow prioritization
        - Pending items notifications
        
        **Performance:** Returns non-paginated results sorted by creation date.
      operationId: getAllWaitingFeedback
      responses:
        '200':
          description: Pending occurrences retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OccurrenceResponseDto'
        '401':
          description: Authentication required
        '403':
          description: Access denied
  /occurrences/feedback:
    post:
      tags:
        - Occurrences
      summary: Submit feedback for an occurrence
      description: |
        Submits staff feedback and review decision for a specific occurrence.
        This action marks the occurrence as reviewed and may trigger student notifications.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Features:**
        - Approval/rejection workflow
        - Optional detailed comments
        - Student association for confirmed occurrences
        - Automatic WebSocket notifications
        - Audit trail for all feedback actions
      operationId: occurrenceFeedback
      parameters:
        - name: occurrenceId
          in: query
          required: true
          description: UUID of the occurrence to provide feedback for
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackDto'
      responses:
        '200':
          description: Feedback submitted successfully and notifications sent
        '400':
          description: Invalid feedback data or validation errors
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Occurrence or specified student not found
        '409':
          description: Conflict - Feedback already exists for this occurrence
  /occurrences/delete:
    delete:
      tags:
        - Occurrences
      summary: Delete an occurrence
      description: |
        Permanently removes an occurrence from the system including associated proof files.
        
        **Required Permissions:** `ROLE_ADMIN` or `ROLE_STAFF`
        
        **Warning:** This action is irreversible. All occurrence data including proof files 
        and feedback will be permanently deleted from both database and file storage.
        
        **Use Cases:**
        - Removing false positives
        - Data cleanup operations
        - Privacy compliance (right to be forgotten)
      operationId: deleteOccurrence
      parameters:
        - name: occurrenceId
          in: query
          required: true
          description: UUID of the occurrence to delete
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Occurrence and associated files deleted successfully
        '401':
          description: Authentication required
        '403':
          description: Access denied
        '404':
          description: Occurrence not found
        '500':
          description: Server error - Failed to delete proof file from storage

  # Charts and Analytics endpoints
...
  /proofs/{fileName}:
    get:
      tags:
        - Proofs
      summary: Retrieve proof file by filename
      description: |
        Retrieves a proof file (evidence) by its filename for authenticated users.
        
        **Required Permissions:** None - Public endpoint (Has to be authenticated)
        
        **Use Cases:**
        - Staff reviewing occurrence evidence
        - Administrative access to proof files
        - Secure file serving with authentication
        
        **Security Features:**
        - Filename validation and sanitization
        - Path traversal protection
        - Content type detection and proper headers
        
        **File Types Supported:** Images, videos, documents, and other multimedia evidence files.
      operationId: getProof
      parameters:
        - name: fileName
          in: path
          required: true
          description: Name of the proof file to retrieve
          schema:
            type: string
            pattern: '^[a-zA-Z0-9._-]+$'
            maxLength: 255
          example: "proof_20240115_143000.jpg"
      responses:
        '200':
          description: Proof file retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Binary content of the proof file
            image/*:
              schema:
                type: string
                format: binary
                description: Image proof file
            video/*:
              schema:
                type: string
                format: binary
                description: Video proof file
            application/pdf:
              schema:
                type: string
                format: binary
                description: PDF document proof file
          headers:
            Content-Type:
              description: MIME type of the file
              schema:
                type: string
                example: "image/jpeg"
            Content-Length:
              description: Size of the file in bytes
              schema:
                type: integer
                example: 2048576
        '401':
          description: Authentication required - Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied - User does not have ADMIN or STAFF role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Proof file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                fileNotFound:
                  summary: File not found error
                  value:
                    timestamp: "2024-01-01T12:00:00.000Z"
                    status: 404
                    error: "Not Found"
                    message: "File not found"
                    path: "/proofs/proof_20240115_143000.jpg"
        '500':
          description: Server error - Failed to read or serve the file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...